+++ inc/define.php
@@ -347,6 +347,13 @@ $CFG_GLPI['itemdevicepci_types']          = ['*'];
 
 $CFG_GLPI['itemdevicesensor_types']       = ['Computer', 'Peripheral'];
 
+$CFG_GLPI['itemdeviceprocessor_types']    = ['Computer'];
+
+$CFG_GLPI['itemdevicesoundcard_types']    = ['Computer'];
+
+$CFG_GLPI['itemdevicegraphiccard_types']  = ['Computer'];
+
+$CFG_GLPI['itemdevicemotherboard_types']  = ['Computer'];
 
 $CFG_GLPI["notificationtemplates_types"]  = ['CartridgeItem', 'Change', 'ConsumableItem',
                                              'Contract', 'Crontask', 'DBConnection',