+++ inc/networkequipment.class.php
@@ -651,6 +651,8 @@ class NetworkEquipment extends CommonDBTM {
 
       $tab = array_merge($tab, Notepad::rawSearchOptionsToAdd());
 
+      $tab = array_merge($tab, Item_Devices::rawSearchOptionsToAdd(get_class($this)));
+
       return $tab;
    }
 }